moved the dashboard section into new docker image directory with docker compose and make target. need to update the argo endpoint configuration through env vars next
