--- dashboard/requirements.txt
@@ -1,1 +1,0 @@
-bettmensch_ai[dashboard]
