+++ docker/dashboard/docker-compose.yaml
@@ -1,0 +1,27 @@
+services:
+  dashboard:
+    build:
+      context: .
+      additional_contexts:
+        image: ./image
+        config: ./config
+        src: ./src
+        sdk: ../../sdk
+      dockerfile: ./Dockerfile
+      args:
+        BASE_IMAGE: python
+        BASE_IMAGE_VERSION: 3.11
+    ports:
+      - 8501:8501
+    command: [streamlit, run, src/0_home.py, --server.port=8501, --server.address=0.0.0.0]
+    network_mode: "host"
+    image: bettmensch88/bettmensch.ai-dashboard:local
+    healthcheck:
+      test: [CMD, curl, -f, http://dashboard:8501/_stcore/health]
+      interval: 10s
+      retries: 5
+      start_period: 5s
+      timeout: 10s
+
+networks:
+  dashboard-net:
