+++ docker/dashboard/src/1_pipelines.py
@@ -1,31 +1,40 @@
 import streamlit as st
 
 st.set_page_config(
-    page_title="Flows", page_icon=":arrow_forward:", layout="wide"
+    page_title="Pipelines",
+    page_icon=":twisted_rightwards_arrows:",
+    layout="wide",
 )
-from typing import Dict, List, Tuple
+from typing import Dict, List, Tuple  # noqa: E402
 
-import pandas as pd
-from bettmensch_ai import DagVisualizationItems, DagVisualizationSettings
-from bettmensch_ai import RegisteredFlow as Flow
-from streamlit_flow import streamlit_flow
-from streamlit_flow.interfaces import StreamlitFlowEdge, StreamlitFlowNode
-from utils import add_logo  # , FLOW_NODE_EMOJI_MAP
-from utils import (
+import pandas as pd  # noqa: E402
+from bettmensch_ai.server import (  # noqa: E402
+    DagVisualizationItems,
+    DagVisualizationSettings,
+)
+from bettmensch_ai.server import RegisteredPipeline as Pipeline  # noqa: E402
+from streamlit_flow import streamlit_flow  # noqa: E402
+from streamlit_flow.interfaces import (  # noqa: E402
+    StreamlitFlowEdge,
+    StreamlitFlowNode,
+)
+from utils import (  # noqa: E402
     PIPELINE_NODE_EMOJI_MAP,
+    add_logo,
     configuration,
     get_colors,
-    get_workflows,
+    get_workflow_templates,
 )
 
 
-def get_flow_meta_data(submitted_flows) -> List[Dict]:
-    """Retrieves the metadata field from all the ArgoWorkflow CRs obtained by
-    `get_workflows`.
+def get_pipeline_meta_data(registered_pipelines) -> List[Dict]:
+    """Retrieves the metadata field from all the ArgoWorkflowTemplate CRs
+    obtained by `get_workflow_templates`.
 
     Args:
-        submitted_flows (List[IoArgoprojWorkflowV1alpha1Workflow]): The
-            ArgoWorkflow CRs obtained by `get_workflows`.
+        registered_pipelines (List[
+            IoArgoprojWorkflowV1alpha1WorkflowTemplate]):
+            The ArgoWorkflowTemplate CRs obtained by `get_workflow_templates`
 
     Returns:
         List[Dict]: A list of dictionaries containing the metadata of each
@@ -33,147 +42,143 @@ def get_flow_meta_data(submitted_flows) -> List[Dict]:
     """
 
     return [
-        submitted_flow.metadata.to_dict() for submitted_flow in submitted_flows
+        registered_pipeline.metadata.to_dict()
+        for registered_pipeline in registered_pipelines
     ]
 
 
-def display_flow_summary_table(formatted_flow_data: Dict) -> pd.DataFrame:
-    """Generates a summary table displaying the key specs of the submitted
-    flows.
+def display_pipeline_summary_table(pipeline_meta_data) -> pd.DataFrame:
+    """Generates a summary table displaying the key specs of the registered
+    pipelines.
 
     Args:
-        flow_meta_data (List[Dict]): The formatted flow data obtained by
-            `get_formatted_flow_data`.
+        pipeline_meta_data (List[Dict]): The pipeline metadata generated by
+            `get_pipeline_meta_data`.
 
     Returns:
-        pd.DataFrame: The flow summary table shown on the frontend.
+        pd.DataFrame: The pipeline summary table shown on the frontend.
     """
 
-    st.markdown("## Submitted flows")
+    st.markdown("## Registered pipelines")
 
-    records = []
-
-    for flow_name in formatted_flow_data["metadata"]:
-        records.append(
-            dict(
-                **formatted_flow_data["state"][flow_name],
-                **formatted_flow_data["metadata"][flow_name],
-            )
-        )
-
-    flow_summary_df = (
-        pd.DataFrame(records)[
-            ["name", "uid", "creation_timestamp", "progress", "phase"]
-        ]
+    pipeline_summary_df = (
+        pd.DataFrame(pipeline_meta_data)[["name", "uid", "creation_timestamp"]]
         .rename(
             columns={
                 "name": "Name",
                 "uid": "ID",
                 "creation_timestamp": "Created",
-                "progress": "Progress",
-                "phase": "Phase",
             }
         )
         .sort_values(by="Created", ignore_index=True)
     )
 
-    st.dataframe(flow_summary_df, hide_index=True)
+    st.dataframe(pipeline_summary_df, hide_index=True)
 
 
-def get_flow_names(flow_meta_data) -> List[str]:
-    """Generates a list of names of all submitted flows.
-
-    Args:
-        flow_meta_data (List[Dict]): The flow metadata generated by
-            `get_flow_meta_data`.
-
-    Returns:
-        List[str]: The names of all available submitted flows.
-    """
-
-    return [resource_meta["name"] for resource_meta in flow_meta_data]
-
-
-def get_formatted_flow_data(submitted_flows, flow_names) -> Dict:
-    """Generates structured flow data for easier frontend useage.
+def get_pipeline_names(pipeline_meta_data) -> List[str]:
+    """Generates a list of names of all registered pipelines.
 
     Args:
-        submitted_flows (List[IoArgoprojWorkflowV1alpha1Workflow]): The
-            ArgoWorkflow CRs obtained by `get_workflows`
-        flow_names (List[str]): The names of flow generated by `get_flow_names`.
+        pipeline_meta_data (List[Dict]): The pipeline metadata generated by
+            `get_pipeline_meta_data`.
 
     Returns:
-        Dict: The formatted flow data.
+        List[str]: The names of all available registered pipelines.
     """
 
-    formatted_flow_data = {
+    return [resource_meta["name"] for resource_meta in pipeline_meta_data]
+
+
+def get_formatted_pipeline_data(registered_pipelines, pipeline_names) -> Dict:
+    """Generates structured pipeline data for easier frontend useage.
+
+    Args:
+        registered_pipelines (List[
+            IoArgoprojWorkflowV1alpha1WorkflowTemplate]):
+            The ArgoWorkflowTemplate CRs obtained by `get_workflow_templates`
+        pipeline_names (List[str]): The names of pipelines generated by
+            `get_pipeline_names`.
+
+    Returns:
+        Dict: The formatted pipeline data.
+    """
+
+    formatted_pipeline_data = {
         "object": {},
         "metadata": {},
-        "state": {},
         "inputs": {},
         "dag": {},
         "templates": {},
     }
 
-    for i, (resource_name, submitted_flow) in enumerate(
-        zip(flow_names, submitted_flows)
+    for i, (resource_name, registered_pipeline) in enumerate(
+        zip(pipeline_names, registered_pipelines)
     ):
         try:
-            flow_dict = Flow.from_argo_workflow_cr(submitted_flow).model_dump()
-            formatted_flow_data["object"][
+            pipeline_dict = Pipeline.from_argo_workflow_cr(
+                registered_pipeline
+            ).model_dump()
+            formatted_pipeline_data["object"][
                 resource_name
-            ] = Flow.from_argo_workflow_cr(submitted_flow)
-            formatted_flow_data["metadata"][resource_name] = flow_dict[
+            ] = Pipeline.from_argo_workflow_cr(registered_pipeline)
+            formatted_pipeline_data["metadata"][resource_name] = pipeline_dict[
                 "metadata"
             ]
-            formatted_flow_data["state"][resource_name] = flow_dict["state"]
-            # formatted_flow_data['artifact_configuration'][resource_name] = flow_dict['artifact_configuration']
-            formatted_flow_data["inputs"][resource_name] = flow_dict["inputs"]
-            formatted_flow_data["dag"][resource_name] = flow_dict["dag"]
-            formatted_flow_data["templates"][resource_name] = flow_dict[
+            formatted_pipeline_data["inputs"][resource_name] = pipeline_dict[
+                "inputs"
+            ]
+            formatted_pipeline_data["dag"][resource_name] = pipeline_dict[
+                "dag"
+            ]  # noqa: E501
+            formatted_pipeline_data["templates"][
+                resource_name
+            ] = pipeline_dict[  # noqa: E501
                 "templates"
             ]
         except Exception as e:
-            raise (e)
             st.write(
-                f"Oops! Could not collect data for Flow {resource_name}: {e}"
-                "Please make sure the Argo Workflow was created with the "
-                "bettmensch.ai SDK and was submitted successfully."
+                f"Oops! Could not collect data for Pipeline {resource_name}: "
+                f"{e} Please make sure the workflow template was created with "
+                "the bettmensch.ai SDK and was submitted successfully."
             )
 
-    return formatted_flow_data
+    return formatted_pipeline_data
 
 
-def display_flow_dropdown(flow_names) -> str:
-    """Display the flow selection dropdown.
+def display_pipeline_dropdown(pipeline_names) -> str:
+    """Display the pipeline selection dropdown.
 
     Args:
-        flow_names (List[str]): The names of flow generated by `get_flow_names`.
+        pipeline_names (List[str]): The names of pipelines generated by
+            `get_pipeline_names`.
 
     Returns:
-        str: The name of the user selected flow.
+        str: The name of the user selected pipeline.
     """
 
-    # display flow selection dropdown
-    selected_flow = st.selectbox(f"Select a Flow:", options=flow_names, index=0)
+    # display pipeline selection dropdown
+    selected_pipeline = st.selectbox(
+        "Select a Pipeline:", options=pipeline_names, index=0
+    )
 
-    return selected_flow
+    return selected_pipeline
 
 
-def display_flow_dag(
-    formatted_flow_data,
-    selected_flow: str,
-    display_flow_ios: bool,
+def display_pipeline_dag(
+    formatted_pipeline_data,
+    selected_pipeline: str,
+    display_pipeline_ios: bool,
     dag_image_height: int,
 ) -> Tuple[DagVisualizationItems, Dict]:
     """_summary_
 
     Args:
-        formatted_flow_data (_type_): The formatted flow data generated by
-            `get_formatted_flow_data`.
-        selected_flow (str): The name of the user selected flow.
-        display_flow_ios (bool): The toggle value of the user selected flow dag
-            I/O detail level in the flow chart.
+        formatted_pipeline_data (_type_): The formatted pipeline data generated
+            by `get_formatted_pipeline_data`.
+        selected_pipeline (str): The name of the user selected pipeline.
+        display_pipeline_ios (bool): The toggle value of the user selected
+            pipeline dag I/O detail level in the flow chart.
         dag_image_height (int): The height of the react flow plugin plot.
 
     Returns:
@@ -181,11 +186,13 @@ def display_flow_dag(
             ReactFlow visualization plugin, and the return of that plugin.
     """
 
-    selected_flow_instance = formatted_flow_data["object"][selected_flow]
+    selected_pipeline_instance = formatted_pipeline_data["object"][
+        selected_pipeline
+    ]
 
     dag_visualization_items = (
-        selected_flow_instance.create_dag_visualization_schema(
-            display_flow_ios,
+        selected_pipeline_instance.create_dag_visualization_schema(
+            display_pipeline_ios,
         )
     )
 
@@ -200,7 +207,9 @@ def display_flow_dag(
         ],
         **DagVisualizationSettings(
             style={
-                "backgroundColor": get_colors("custom").secondaryBackgroundColor
+                "backgroundColor": get_colors(
+                    "custom"
+                ).secondaryBackgroundColor  # noqa: E501
             },
             height=dag_image_height,
         ).model_dump(),
@@ -209,20 +218,21 @@ def display_flow_dag(
     return dag_visualization_items, dag_visualization_element
 
 
-def display_flow_dag_selection(
-    formatted_flow_data: Dict,
-    selected_flow: str,
+def display_pipeline_dag_selection(
+    formatted_pipeline_data: Dict,
+    selected_pipeline: str,
     dag_visualization_element,
     tab_container_height: int,
 ):
-    """Generates a tabbed, in depth view of the user selected DAG task node element.
+    """Generates a tabbed, in depth view of the user selected DAG task node
+    element.
 
     Args:
-        formatted_flow_data (Dict): The formatted flow data as generated by
-            `get_formatted_flow_data`
-        selected_flow (str): The user selected flow name.
+        formatted_pipeline_data (Dict): The formatted pipeline data as
+            generated by `get_formatted_pipeline_data`
+        selected_pipeline (str): The user selected pipeline name.
         dag_visualization_element (_type_): The user selected element from the
-            DAG visuals, as returned by `display_flow_dag`
+            DAG visuals, as returned by `display_pipeline_dag`
     """
 
     try:
@@ -234,26 +244,17 @@ def display_flow_dag_selection(
 
         if element_is_task_node:
             st.markdown(
-                f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: `{element_id}`"
+                f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: `{element_id}`"  # noqa: E501
             )
-            (
-                task_inputs_tab,
-                task_outputs_tab,
-                task_script_tab,
-                task_state_tab,
-            ) = st.tabs(
-                [
-                    "Component Inputs",
-                    "Component Outputs",
-                    "Component Script",
-                    "Component State",
-                ]
+            task_inputs_tab, task_outputs_tab, task_script_tab = st.tabs(
+                ["Component Inputs", "Component Outputs", "Component Script"]
             )
-            flow = formatted_flow_data["object"][selected_flow]
-            task = flow.get_dag_task(element_id).model_dump()
+            pipeline = formatted_pipeline_data["object"][selected_pipeline]
+            task = pipeline.get_dag_task(element_id).model_dump()
 
             with task_inputs_tab:
                 with st.container(height=tab_container_height, border=False):
+                    # build task input parameters table
                     if task["inputs"]["parameters"]:
                         task_inputs_parameters_df = pd.DataFrame(
                             task["inputs"]["parameters"]
@@ -271,7 +272,7 @@ def display_flow_dag_selection(
                         ).rename(
                             columns={
                                 "name": "Name",
-                                "value": "Value",
+                                "value": "Default",
                                 "node": "Upstream Task",
                                 "output_name": "Upstream Output",
                                 "output_type": "Upstream Type",
@@ -281,6 +282,7 @@ def display_flow_dag_selection(
                     else:
                         task_inputs_parameters_formatted_df = pd.DataFrame()
 
+                    # build task input artifact table
                     if task["inputs"]["artifacts"]:
                         task_inputs_artifacts_df = pd.DataFrame(
                             task["inputs"]["artifacts"]
@@ -298,7 +300,6 @@ def display_flow_dag_selection(
                         ).rename(
                             columns={
                                 "name": "Name",
-                                "s3_prefix": "S3 Prefix",
                                 "node": "Upstream Task",
                                 "output_name": "Upstream Output",
                                 "output_type": "Upstream Type",
@@ -319,7 +320,7 @@ def display_flow_dag_selection(
 
             with task_outputs_tab:
                 with st.container(height=tab_container_height, border=False):
-
+                    # build task output parameters table
                     if task["outputs"]["parameters"]:
                         task_outputs_parameters_df = pd.DataFrame(
                             task["outputs"]["parameters"]
@@ -330,7 +331,6 @@ def display_flow_dag_selection(
                             ).rename(
                                 columns={
                                     "name": "Name",
-                                    "value": "Value",
                                 },
                                 inplace=False,
                             )
@@ -338,6 +338,7 @@ def display_flow_dag_selection(
                     else:
                         task_outputs_parameters_formatted_df = pd.DataFrame()
 
+                    # build task output artifact table
                     if task["outputs"]["artifacts"]:
                         task_outputs_artifacts_df = pd.DataFrame(
                             task["outputs"]["artifacts"]
@@ -348,7 +349,6 @@ def display_flow_dag_selection(
                             ).rename(
                                 columns={
                                     "name": "Name",
-                                    "s3_prefix": "S3 Prefix",
                                 },
                                 inplace=False,
                             )
@@ -368,53 +368,20 @@ def display_flow_dag_selection(
             with task_script_tab:
                 with st.container(height=tab_container_height, border=False):
                     st.json(
-                        flow.get_template(task["template"]).model_dump()[
+                        pipeline.get_template(task["template"]).model_dump()[
                             "script"
                         ]
                     )
-
-            with task_state_tab:
-                with st.container(height=tab_container_height, border=False):
-                    task_state = dict(
-                        [
-                            (k, v)
-                            for k, v in task.items()
-                            if k
-                            in (
-                                "id",
-                                "pod_name",
-                                "host_node_name",
-                                "phase",
-                                "logs",
-                            )
-                        ]
-                    )
-                    task_state_formatted_df = (
-                        pd.DataFrame(task_state, index=[0])
-                        .rename(
-                            columns={
-                                "id": "Flow Task ID",
-                                "pod_name": "Pod",
-                                "host_node_name": "Host Node",
-                                "phase": "Phase",
-                                "logs": "Logs",
-                            },
-                            inplace=False,
-                        )
-                        .T
-                    )
-                    st.dataframe(task_state_formatted_df, hide_index=False)
-
         else:
             st.markdown(
                 f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: None"
             )
             st.write(
-                "Select a `Component` by clicking on the corresponding "
+                "Select a component by clicking on the corresponding "
                 f"{PIPELINE_NODE_EMOJI_MAP['task']} node."
             )
 
-    except TypeError as e:
+    except TypeError:
         st.markdown(f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: None")
         st.write(
             "Select a `Component` by clicking on the corresponding "
@@ -422,62 +389,72 @@ def display_flow_dag_selection(
         )
 
 
-def display_selected_flow(
-    formatted_flow_data,
-    selected_flow,
+def display_selected_pipeline(
+    formatted_pipeline_data,
+    selected_pipeline,
     tab_container_height: int = 420,
     dag_image_height: int = 1100,
 ):
     """Utility to display DAG flow chart and all relevant specs in tabbed
-    layout for a user selected flow.
+    layout for a user selected pipeline.
 
     Args:
-        formatted_flow_data (Dict): The formatted flow data as obtained by
-            `get_formatted_flow_data`.
-        selected_flow (str): The name of the user selected flow.
+        formatted_pipeline_data (Dict): The formatted pipeline data as obtained
+            by `get_formatted_pipeline_data`.
+        selected_pipeline (str): The name of the user selected pipeline.
     """
 
     dag_col, spec_col = st.columns([3, 2])
 
     with dag_col:
-        display_pipeline_ios = st.toggle(f"Display flow & component I/O")
-        dag_visualization_schema, dag_visualization_element = display_flow_dag(
-            formatted_flow_data,
-            selected_flow,
+        display_pipeline_ios = st.toggle("Display pipeline & component I/O")
+        (
+            dag_visualization_schema,
+            dag_visualization_element,
+        ) = display_pipeline_dag(
+            formatted_pipeline_data,
+            selected_pipeline,
             display_pipeline_ios,
             dag_image_height,
         )
 
-    # display flow level data
+    # display pipeline level data
     with spec_col:
-        st.markdown(f"### :arrow_forward: Flow: `{selected_flow}`")
+        st.markdown(
+            "### :twisted_rightwards_arrows: Pipeline: `"
+            f"{selected_pipeline}`"
+        )
 
         tab_inputs, tab_metadata, tab_dag, tab_templates = st.tabs(
             [
-                "Flow Inputs",
-                "Flow Meta Data",
-                "Flow DAG",
-                "Flow Templates",
+                "Pipeline Inputs",
+                "Pipeline Meta Data",
+                "Pipeline DAG",
+                "Pipeline Templates",
             ]
         )
 
         with tab_inputs:
             with st.container(height=tab_container_height, border=False):
-                flow_inputs = formatted_flow_data["inputs"][selected_flow]
-                flow_inputs_formatted_df = pd.DataFrame(flow_inputs).rename(
+                pipeline_inputs = formatted_pipeline_data["inputs"][
+                    selected_pipeline
+                ]
+                pipeline_inputs_formatted_df = pd.DataFrame(
+                    pipeline_inputs
+                ).rename(
                     columns={
                         "name": "Name",
-                        "value": "Value",
+                        "value": "Default",
                     }
                 )
                 st.write(":page_with_curl: Parameters")
-                st.dataframe(flow_inputs_formatted_df, hide_index=True)
+                st.dataframe(pipeline_inputs_formatted_df, hide_index=True)
 
         with tab_metadata:
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_flow_data["metadata"][selected_flow],
+                    formatted_pipeline_data["metadata"][selected_pipeline],
                     expanded=True,
                 )
 
@@ -485,21 +462,22 @@ def display_selected_flow(
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_flow_data["dag"][selected_flow], expanded=True
+                    formatted_pipeline_data["dag"][selected_pipeline],
+                    expanded=True,
                 )
 
         with tab_templates:
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_flow_data["templates"][selected_flow],
+                    formatted_pipeline_data["templates"][selected_pipeline],
                     expanded=True,
                 )
 
         # display task level data
-        display_flow_dag_selection(
-            formatted_flow_data,
-            selected_flow,
+        display_pipeline_dag_selection(
+            formatted_pipeline_data,
+            selected_pipeline,
             dag_visualization_element,
             tab_container_height,
         )
@@ -512,27 +490,30 @@ def main():
 
     st.markdown(
         """
-        # :arrow_forward: Flows
-        
-        A `Flow` is the *execution* of a workflow, i.e. it is the running of a `Pipeline`.
+        # :twisted_rightwards_arrows: Pipelines
+
+        A `Pipeline` is the *definition* of a workflow, i.e. it describes a DAG
+        declaring the logic and dependencies that will be executd at runtime.
         """
     )
 
-    workflows = get_workflows(configuration)
+    workflow_templates = get_workflow_templates(configuration)
 
-    meta_data = get_flow_meta_data(workflows)
+    meta_data = get_pipeline_meta_data(workflow_templates)
+    display_pipeline_summary_table(meta_data)
 
-    names = get_flow_names(meta_data)
+    names = get_pipeline_names(meta_data)
 
-    formatted_flow_data = get_formatted_flow_data(workflows, names)
-
-    display_flow_summary_table(formatted_flow_data)
-
-    selected_flow = display_flow_dropdown(names)
-
-    dag_visualization_schema, dag_visualization_element = display_selected_flow(
-        formatted_flow_data, selected_flow
+    formatted_pipeline_data = get_formatted_pipeline_data(
+        workflow_templates, names
     )
+
+    selected_pipeline = display_pipeline_dropdown(names)
+
+    (
+        dag_visualization_schema,
+        dag_visualization_element,
+    ) = display_selected_pipeline(formatted_pipeline_data, selected_pipeline)
 
     with st.sidebar:
         add_logo(sidebar=True)
