+++ docker/dashboard/src/2_flows.py
@@ -1,33 +1,37 @@
 import streamlit as st
 
 st.set_page_config(
-    page_title="Pipelines",
-    page_icon=":twisted_rightwards_arrows:",
-    layout="wide",
+    page_title="Flows", page_icon=":arrow_forward:", layout="wide"
 )
-from typing import Dict, List, Tuple
+from typing import Dict, List, Tuple  # noqa: E402
 
-import pandas as pd
-from bettmensch_ai import DagVisualizationItems, DagVisualizationSettings
-from bettmensch_ai import RegisteredPipeline as Pipeline
-from streamlit_flow import streamlit_flow
-from streamlit_flow.interfaces import StreamlitFlowEdge, StreamlitFlowNode
-from utils import (
+import pandas as pd  # noqa: E402
+from bettmensch_ai.server import (  # noqa: E402,E501
+    DagVisualizationItems,
+    DagVisualizationSettings,
+)
+from bettmensch_ai.server import RegisteredFlow as Flow  # noqa: E402
+from streamlit_flow import streamlit_flow  # noqa: E402
+from streamlit_flow.interfaces import (  # noqa: E402
+    StreamlitFlowEdge,
+    StreamlitFlowNode,
+)
+from utils import add_logo  # , FLOW_NODE_EMOJI_MAP # noqa: E402
+from utils import (  # noqa: E402
     PIPELINE_NODE_EMOJI_MAP,
-    add_logo,
     configuration,
     get_colors,
-    get_workflow_templates,
+    get_workflows,
 )
 
 
-def get_pipeline_meta_data(registered_pipelines) -> List[Dict]:
-    """Retrieves the metadata field from all the ArgoWorkflowTemplate CRs
-    obtained by `get_workflow_templates`.
+def get_flow_meta_data(submitted_flows) -> List[Dict]:
+    """Retrieves the metadata field from all the ArgoWorkflow CRs obtained by
+    `get_workflows`.
 
     Args:
-        registered_pipelines (List[IoArgoprojWorkflowV1alpha1WorkflowTemplate]):
-            The ArgoWorkflowTemplate CRs obtained by `get_workflow_templates`
+        submitted_flows (List[IoArgoprojWorkflowV1alpha1Workflow]): The
+            ArgoWorkflow CRs obtained by `get_workflows`.
 
     Returns:
         List[Dict]: A list of dictionaries containing the metadata of each
@@ -35,138 +39,148 @@ def get_pipeline_meta_data(registered_pipelines) -> List[Dict]:
     """
 
     return [
-        registered_pipeline.metadata.to_dict()
-        for registered_pipeline in registered_pipelines
+        submitted_flow.metadata.to_dict() for submitted_flow in submitted_flows
     ]
 
 
-def display_pipeline_summary_table(pipeline_meta_data) -> pd.DataFrame:
-    """Generates a summary table displaying the key specs of the registered
-    pipelines.
+def display_flow_summary_table(formatted_flow_data: Dict) -> pd.DataFrame:
+    """Generates a summary table displaying the key specs of the submitted
+    flows.
 
     Args:
-        pipeline_meta_data (List[Dict]): The pipeline metadata generated by
-            `get_pipeline_meta_data`.
+        flow_meta_data (List[Dict]): The formatted flow data obtained by
+            `get_formatted_flow_data`.
 
     Returns:
-        pd.DataFrame: The pipeline summary table shown on the frontend.
+        pd.DataFrame: The flow summary table shown on the frontend.
     """
 
-    st.markdown("## Registered pipelines")
+    st.markdown("## Submitted flows")
 
-    pipeline_summary_df = (
-        pd.DataFrame(pipeline_meta_data)[["name", "uid", "creation_timestamp"]]
+    records = []
+
+    for flow_name in formatted_flow_data["metadata"]:
+        records.append(
+            dict(
+                **formatted_flow_data["state"][flow_name],
+                **formatted_flow_data["metadata"][flow_name],
+            )
+        )
+
+    flow_summary_df = (
+        pd.DataFrame(records)[
+            ["name", "uid", "creation_timestamp", "progress", "phase"]
+        ]
         .rename(
             columns={
                 "name": "Name",
                 "uid": "ID",
                 "creation_timestamp": "Created",
+                "progress": "Progress",
+                "phase": "Phase",
             }
         )
         .sort_values(by="Created", ignore_index=True)
     )
 
-    st.dataframe(pipeline_summary_df, hide_index=True)
+    st.dataframe(flow_summary_df, hide_index=True)
 
 
-def get_pipeline_names(pipeline_meta_data) -> List[str]:
-    """Generates a list of names of all registered pipelines.
-
-    Args:
-        pipeline_meta_data (List[Dict]): The pipeline metadata generated by
-            `get_pipeline_meta_data`.
-
-    Returns:
-        List[str]: The names of all available registered pipelines.
-    """
-
-    return [resource_meta["name"] for resource_meta in pipeline_meta_data]
-
-
-def get_formatted_pipeline_data(registered_pipelines, pipeline_names) -> Dict:
-    """Generates structured pipeline data for easier frontend useage.
+def get_flow_names(flow_meta_data) -> List[str]:
+    """Generates a list of names of all submitted flows.
 
     Args:
-        registered_pipelines (List[IoArgoprojWorkflowV1alpha1WorkflowTemplate]):
-            The ArgoWorkflowTemplate CRs obtained by `get_workflow_templates`
-        pipeline_names (List[str]): The names of pipelines generated by
-            `get_pipeline_names`.
+        flow_meta_data (List[Dict]): The flow metadata generated by
+            `get_flow_meta_data`.
 
     Returns:
-        Dict: The formatted pipeline data.
+        List[str]: The names of all available submitted flows.
     """
 
-    formatted_pipeline_data = {
+    return [resource_meta["name"] for resource_meta in flow_meta_data]
+
+
+def get_formatted_flow_data(submitted_flows, flow_names) -> Dict:
+    """Generates structured flow data for easier frontend useage.
+
+    Args:
+        submitted_flows (List[IoArgoprojWorkflowV1alpha1Workflow]): The
+            ArgoWorkflow CRs obtained by `get_workflows`
+        flow_names (List[str]): The names of flow generated by
+            `get_flow_names`.
+
+    Returns:
+        Dict: The formatted flow data.
+    """
+
+    formatted_flow_data = {
         "object": {},
         "metadata": {},
+        "state": {},
         "inputs": {},
         "dag": {},
         "templates": {},
     }
 
-    for i, (resource_name, registered_pipeline) in enumerate(
-        zip(pipeline_names, registered_pipelines)
+    for i, (resource_name, submitted_flow) in enumerate(
+        zip(flow_names, submitted_flows)
     ):
         try:
-            pipeline_dict = Pipeline.from_argo_workflow_cr(
-                registered_pipeline
-            ).model_dump()
-            formatted_pipeline_data["object"][
+            flow_dict = Flow.from_argo_workflow_cr(submitted_flow).model_dump()
+            formatted_flow_data["object"][
                 resource_name
-            ] = Pipeline.from_argo_workflow_cr(registered_pipeline)
-            formatted_pipeline_data["metadata"][resource_name] = pipeline_dict[
+            ] = Flow.from_argo_workflow_cr(submitted_flow)
+            formatted_flow_data["metadata"][resource_name] = flow_dict[
                 "metadata"
             ]
-            formatted_pipeline_data["inputs"][resource_name] = pipeline_dict[
-                "inputs"
-            ]
-            formatted_pipeline_data["dag"][resource_name] = pipeline_dict["dag"]
-            formatted_pipeline_data["templates"][resource_name] = pipeline_dict[
+            formatted_flow_data["state"][resource_name] = flow_dict["state"]
+            formatted_flow_data["inputs"][resource_name] = flow_dict["inputs"]
+            formatted_flow_data["dag"][resource_name] = flow_dict["dag"]
+            formatted_flow_data["templates"][resource_name] = flow_dict[
                 "templates"
             ]
         except Exception as e:
+            raise (e)
             st.write(
-                f"Oops! Could not collect data for Pipeline {resource_name}: "
-                f"{e} Please make sure the workflow template was created with "
-                "the bettmensch.ai SDK and was submitted successfully."
+                f"Oops! Could not collect data for Flow {resource_name}: {e}"
+                "Please make sure the Argo Workflow was created with the "
+                "bettmensch.ai SDK and was submitted successfully."
             )
 
-    return formatted_pipeline_data
+    return formatted_flow_data
 
 
-def display_pipeline_dropdown(pipeline_names) -> str:
-    """Display the pipeline selection dropdown.
+def display_flow_dropdown(flow_names) -> str:
+    """Display the flow selection dropdown.
 
     Args:
-        pipeline_names (List[str]): The names of pipelines generated by
-            `get_pipeline_names`.
+        flow_names (List[str]): The names of flow generated by
+            `get_flow_names`.
 
     Returns:
-        str: The name of the user selected pipeline.
+        str: The name of the user selected flow.
     """
 
-    # display pipeline selection dropdown
-    selected_pipeline = st.selectbox(
-        f"Select a Pipeline:", options=pipeline_names, index=0
-    )
+    # display flow selection dropdown
+    selected_flow = st.selectbox("Select a Flow:", options=flow_names, index=0)
 
-    return selected_pipeline
+    return selected_flow
 
 
-def display_pipeline_dag(
-    formatted_pipeline_data,
-    selected_pipeline: str,
-    display_pipeline_ios: bool,
+def display_flow_dag(
+    formatted_flow_data,
+    selected_flow: str,
+    display_flow_ios: bool,
     dag_image_height: int,
 ) -> Tuple[DagVisualizationItems, Dict]:
     """_summary_
 
     Args:
-        formatted_pipeline_data (_type_): The formatted pipeline data generated
-            by `get_formatted_pipeline_data`.
-        selected_pipeline (str): The name of the user selected pipeline.
-        display_pipeline_ios (bool): The toggle value of the user selected
-            pipeline dag I/O detail level in the flow chart.
+        formatted_flow_data (_type_): The formatted flow data generated by
+            `get_formatted_flow_data`.
+        selected_flow (str): The name of the user selected flow.
+        display_flow_ios (bool): The toggle value of the user selected flow dag
+            I/O detail level in the flow chart.
         dag_image_height (int): The height of the react flow plugin plot.
 
     Returns:
@@ -174,13 +188,11 @@ def display_pipeline_dag(
             ReactFlow visualization plugin, and the return of that plugin.
     """
 
-    selected_pipeline_instance = formatted_pipeline_data["object"][
-        selected_pipeline
-    ]
+    selected_flow_instance = formatted_flow_data["object"][selected_flow]
 
     dag_visualization_items = (
-        selected_pipeline_instance.create_dag_visualization_schema(
-            display_pipeline_ios,
+        selected_flow_instance.create_dag_visualization_schema(
+            display_flow_ios,
         )
     )
 
@@ -195,7 +207,9 @@ def display_pipeline_dag(
         ],
         **DagVisualizationSettings(
             style={
-                "backgroundColor": get_colors("custom").secondaryBackgroundColor
+                "backgroundColor": get_colors(
+                    "custom"
+                ).secondaryBackgroundColor  # noqa: E501
             },
             height=dag_image_height,
         ).model_dump(),
@@ -204,9 +218,9 @@ def display_pipeline_dag(
     return dag_visualization_items, dag_visualization_element
 
 
-def display_pipeline_dag_selection(
-    formatted_pipeline_data: Dict,
-    selected_pipeline: str,
+def display_flow_dag_selection(
+    formatted_flow_data: Dict,
+    selected_flow: str,
     dag_visualization_element,
     tab_container_height: int,
 ):
@@ -214,11 +228,11 @@ def display_pipeline_dag_selection(
     element.
 
     Args:
-        formatted_pipeline_data (Dict): The formatted pipeline data as
-            generated by `get_formatted_pipeline_data`
-        selected_pipeline (str): The user selected pipeline name.
+        formatted_flow_data (Dict): The formatted flow data as generated by
+            `get_formatted_flow_data`
+        selected_flow (str): The user selected flow name.
         dag_visualization_element (_type_): The user selected element from the
-            DAG visuals, as returned by `display_pipeline_dag`
+            DAG visuals, as returned by `display_flow_dag`
     """
 
     try:
@@ -230,17 +244,26 @@ def display_pipeline_dag_selection(
 
         if element_is_task_node:
             st.markdown(
-                f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: `{element_id}`"
+                f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: `{element_id}`"  # noqa: E501
             )
-            task_inputs_tab, task_outputs_tab, task_script_tab = st.tabs(
-                ["Component Inputs", "Component Outputs", "Component Script"]
+            (
+                task_inputs_tab,
+                task_outputs_tab,
+                task_script_tab,
+                task_state_tab,
+            ) = st.tabs(
+                [
+                    "Component Inputs",
+                    "Component Outputs",
+                    "Component Script",
+                    "Component State",
+                ]
             )
-            pipeline = formatted_pipeline_data["object"][selected_pipeline]
-            task = pipeline.get_dag_task(element_id).model_dump()
+            flow = formatted_flow_data["object"][selected_flow]
+            task = flow.get_dag_task(element_id).model_dump()
 
             with task_inputs_tab:
                 with st.container(height=tab_container_height, border=False):
-                    # build task input parameters table
                     if task["inputs"]["parameters"]:
                         task_inputs_parameters_df = pd.DataFrame(
                             task["inputs"]["parameters"]
@@ -258,7 +281,7 @@ def display_pipeline_dag_selection(
                         ).rename(
                             columns={
                                 "name": "Name",
-                                "value": "Default",
+                                "value": "Value",
                                 "node": "Upstream Task",
                                 "output_name": "Upstream Output",
                                 "output_type": "Upstream Type",
@@ -268,7 +291,6 @@ def display_pipeline_dag_selection(
                     else:
                         task_inputs_parameters_formatted_df = pd.DataFrame()
 
-                    # build task input artifact table
                     if task["inputs"]["artifacts"]:
                         task_inputs_artifacts_df = pd.DataFrame(
                             task["inputs"]["artifacts"]
@@ -286,6 +308,7 @@ def display_pipeline_dag_selection(
                         ).rename(
                             columns={
                                 "name": "Name",
+                                "s3_prefix": "S3 Prefix",
                                 "node": "Upstream Task",
                                 "output_name": "Upstream Output",
                                 "output_type": "Upstream Type",
@@ -306,7 +329,7 @@ def display_pipeline_dag_selection(
 
             with task_outputs_tab:
                 with st.container(height=tab_container_height, border=False):
-                    # build task output parameters table
+
                     if task["outputs"]["parameters"]:
                         task_outputs_parameters_df = pd.DataFrame(
                             task["outputs"]["parameters"]
@@ -317,6 +340,7 @@ def display_pipeline_dag_selection(
                             ).rename(
                                 columns={
                                     "name": "Name",
+                                    "value": "Value",
                                 },
                                 inplace=False,
                             )
@@ -324,7 +348,6 @@ def display_pipeline_dag_selection(
                     else:
                         task_outputs_parameters_formatted_df = pd.DataFrame()
 
-                    # build task output artifact table
                     if task["outputs"]["artifacts"]:
                         task_outputs_artifacts_df = pd.DataFrame(
                             task["outputs"]["artifacts"]
@@ -335,6 +358,7 @@ def display_pipeline_dag_selection(
                             ).rename(
                                 columns={
                                     "name": "Name",
+                                    "s3_prefix": "S3 Prefix",
                                 },
                                 inplace=False,
                             )
@@ -354,20 +378,53 @@ def display_pipeline_dag_selection(
             with task_script_tab:
                 with st.container(height=tab_container_height, border=False):
                     st.json(
-                        pipeline.get_template(task["template"]).model_dump()[
+                        flow.get_template(task["template"]).model_dump()[
                             "script"
                         ]
                     )
+
+            with task_state_tab:
+                with st.container(height=tab_container_height, border=False):
+                    task_state = dict(
+                        [
+                            (k, v)
+                            for k, v in task.items()
+                            if k
+                            in (
+                                "id",
+                                "pod_name",
+                                "host_node_name",
+                                "phase",
+                                "logs",
+                            )
+                        ]
+                    )
+                    task_state_formatted_df = (
+                        pd.DataFrame(task_state, index=[0])
+                        .rename(
+                            columns={
+                                "id": "Flow Task ID",
+                                "pod_name": "Pod",
+                                "host_node_name": "Host Node",
+                                "phase": "Phase",
+                                "logs": "Logs",
+                            },
+                            inplace=False,
+                        )
+                        .T
+                    )
+                    st.dataframe(task_state_formatted_df, hide_index=False)
+
         else:
             st.markdown(
                 f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: None"
             )
             st.write(
-                "Select a component by clicking on the corresponding "
+                "Select a `Component` by clicking on the corresponding "
                 f"{PIPELINE_NODE_EMOJI_MAP['task']} node."
             )
 
-    except TypeError as e:
+    except TypeError:
         st.markdown(f"### {PIPELINE_NODE_EMOJI_MAP['task']} Component: None")
         st.write(
             "Select a `Component` by clicking on the corresponding "
@@ -375,72 +432,62 @@ def display_pipeline_dag_selection(
         )
 
 
-def display_selected_pipeline(
-    formatted_pipeline_data,
-    selected_pipeline,
+def display_selected_flow(
+    formatted_flow_data,
+    selected_flow,
     tab_container_height: int = 420,
     dag_image_height: int = 1100,
 ):
     """Utility to display DAG flow chart and all relevant specs in tabbed
-    layout for a user selected pipeline.
+    layout for a user selected flow.
 
     Args:
-        formatted_pipeline_data (Dict): The formatted pipeline data as obtained
-            by `get_formatted_pipeline_data`.
-        selected_pipeline (str): The name of the user selected pipeline.
+        formatted_flow_data (Dict): The formatted flow data as obtained by
+            `get_formatted_flow_data`.
+        selected_flow (str): The name of the user selected flow.
     """
 
     dag_col, spec_col = st.columns([3, 2])
 
     with dag_col:
-        display_pipeline_ios = st.toggle(f"Display pipeline & component I/O")
-        (
-            dag_visualization_schema,
-            dag_visualization_element,
-        ) = display_pipeline_dag(
-            formatted_pipeline_data,
-            selected_pipeline,
+        display_pipeline_ios = st.toggle("Display flow & component I/O")
+        dag_visualization_schema, dag_visualization_element = display_flow_dag(
+            formatted_flow_data,
+            selected_flow,
             display_pipeline_ios,
             dag_image_height,
         )
 
-    # display pipeline level data
+    # display flow level data
     with spec_col:
-        st.markdown(
-            "### :twisted_rightwards_arrows: Pipeline: `"
-            f"{selected_pipeline}`"
-        )
+        st.markdown(f"### :arrow_forward: Flow: `{selected_flow}`")
 
         tab_inputs, tab_metadata, tab_dag, tab_templates = st.tabs(
             [
-                "Pipeline Inputs",
-                "Pipeline Meta Data",
-                "Pipeline DAG",
-                "Pipeline Templates",
+                "Flow Inputs",
+                "Flow Meta Data",
+                "Flow DAG",
+                "Flow Templates",
             ]
         )
 
         with tab_inputs:
             with st.container(height=tab_container_height, border=False):
-                pipeline_inputs = formatted_pipeline_data["inputs"][
-                    selected_pipeline
-                ]
-                pipeline_inputs_formatted_df = pd.DataFrame(
-                    pipeline_inputs
-                ).rename(
+                flow_inputs = formatted_flow_data["inputs"][selected_flow]
+                flow_inputs_formatted_df = pd.DataFrame(flow_inputs).rename(
                     columns={
                         "name": "Name",
-                        "value": "Default",
+                        "value": "Value",
                     }
                 )
                 st.write(":page_with_curl: Parameters")
-                st.dataframe(pipeline_inputs_formatted_df, hide_index=True)
+                st.dataframe(flow_inputs_formatted_df, hide_index=True)
 
         with tab_metadata:
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_pipeline_data["metadata"][selected_pipeline],
+                    formatted_flow_data["metadata"][selected_flow],
                     expanded=True,
                 )
 
@@ -448,22 +495,21 @@ def display_selected_pipeline(
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_pipeline_data["dag"][selected_pipeline],
-                    expanded=True,
+                    formatted_flow_data["dag"][selected_flow], expanded=True
                 )
 
         with tab_templates:
             with st.container(height=tab_container_height, border=False):
                 st.markdown("### Spec")
                 st.json(
-                    formatted_pipeline_data["templates"][selected_pipeline],
+                    formatted_flow_data["templates"][selected_flow],
                     expanded=True,
                 )
 
         # display task level data
-        display_pipeline_dag_selection(
-            formatted_pipeline_data,
-            selected_pipeline,
+        display_flow_dag_selection(
+            formatted_flow_data,
+            selected_flow,
             dag_visualization_element,
             tab_container_height,
         )
@@ -476,29 +522,26 @@ def main():
 
     st.markdown(
         """
-        # :twisted_rightwards_arrows: Pipelines
-        
-        A `Pipeline` is the *definition* of a workflow, i.e. it describes a DAG declaring the logic and dependencies that will be executd at runtime.
+        # :arrow_forward: Flows
+
+        A `Flow` is the *execution* of a workflow, i.e. it is the running of a
+        `Pipeline`.
         """
     )
 
-    workflow_templates = get_workflow_templates(configuration)
+    workflows = get_workflows(configuration)
 
-    meta_data = get_pipeline_meta_data(workflow_templates)
-    display_pipeline_summary_table(meta_data)
+    meta_data = get_flow_meta_data(workflows)
 
-    names = get_pipeline_names(meta_data)
+    names = get_flow_names(meta_data)
 
-    formatted_pipeline_data = get_formatted_pipeline_data(
-        workflow_templates, names
-    )
+    formatted_flow_data = get_formatted_flow_data(workflows, names)
 
-    selected_pipeline = display_pipeline_dropdown(names)
+    display_flow_summary_table(formatted_flow_data)
 
-    (
-        dag_visualization_schema,
-        dag_visualization_element,
-    ) = display_selected_pipeline(formatted_pipeline_data, selected_pipeline)
+    selected_flow = display_flow_dropdown(names)
+
+    _, _ = display_selected_flow(formatted_flow_data, selected_flow)
 
     with st.sidebar:
         add_logo(sidebar=True)
