+++ makefile
@@ -1,1 +1,2 @@
 include docker/component/makefile
+include docker/dashboard/makefile
